+++ familyhelper-life-impl/src/main/java/com/dwarfeng/familyhelper/life/impl/handler/PushHandlerImpl.java
@@ -1,0 +1,45 @@
+package com.dwarfeng.familyhelper.life.impl.handler;
+
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityGeneratedByDriverPushInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityRemindedByDriverPushInfo;
+import com.dwarfeng.familyhelper.life.stack.handler.PushHandler;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+
+@Component
+public class PushHandlerImpl implements PushHandler {
+
+    private final List<Pusher> pushers;
+
+    @Value("${pusher.type}")
+    private String pusherType;
+
+    private Pusher pusher;
+
+    public PushHandlerImpl(List<Pusher> pushers) {
+        this.pushers = Optional.ofNullable(pushers).orElse(Collections.emptyList());
+    }
+
+    @PostConstruct
+    public void init() throws HandlerException {
+        this.pusher = pushers.stream().filter(p -> p.supportType(pusherType)).findAny().orElseThrow(
+                () -> new HandlerException("未知的 pusher 类型: " + pusherType)
+        );
+    }
+
+    @Override
+    public void activityGeneratedByDriver(ActivityGeneratedByDriverPushInfo pushInfo) throws HandlerException {
+        pusher.activityGeneratedByDriver(pushInfo);
+    }
+
+    @Override
+    public void activityRemindedByDriver(ActivityRemindedByDriverPushInfo pushInfo) throws HandlerException {
+        pusher.activityRemindedByDriver(pushInfo);
+    }
+}