+++ familyhelper-life-impl/src/main/java/com/dwarfeng/familyhelper/life/impl/handler/pusher/MultiPusher.java
@@ -1,0 +1,79 @@
+package com.dwarfeng.familyhelper.life.impl.handler.pusher;
+
+import com.dwarfeng.familyhelper.life.impl.handler.Pusher;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityGeneratedByDriverPushInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityRemindedByDriverPushInfo;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import java.util.*;
+
+/**
+ * 同时将消息推送给所有代理的多重推送器。
+ *
+ * @author DwArFeng
+ * @since 1.1.1
+ */
+@Component
+public class MultiPusher extends AbstractPusher {
+
+    public static final String PUSHER_TYPE = "multi";
+    private static final Logger LOGGER = LoggerFactory.getLogger(MultiPusher.class);
+
+    private final List<Pusher> pushers;
+
+    @Value("${pusher.multi.delegate_types}")
+    private String delegateTypes;
+
+    private final List<Pusher> delegates = new ArrayList<>();
+
+    public MultiPusher(@Lazy List<Pusher> pushers) {
+        super(PUSHER_TYPE);
+        this.pushers = Optional.ofNullable(pushers).orElse(Collections.emptyList());
+    }
+
+    @PostConstruct
+    public void init() throws HandlerException {
+        StringTokenizer st = new StringTokenizer(delegateTypes, ",");
+        while (st.hasMoreTokens()) {
+            String delegateType = st.nextToken();
+            delegates.add(pushers.stream().filter(p -> p.supportType(delegateType)).findAny()
+                    .orElseThrow(() -> new HandlerException("未知的 pusher 类型: " + delegateType)));
+        }
+    }
+
+    @Override
+    public void activityGeneratedByDriver(ActivityGeneratedByDriverPushInfo pushInfo) {
+        for (Pusher delegate : delegates) {
+            try {
+                delegate.activityGeneratedByDriver(pushInfo);
+            } catch (Exception e) {
+                LOGGER.warn("代理推送器推送消息失败，异常信息如下: ", e);
+            }
+        }
+    }
+
+    @Override
+    public void activityRemindedByDriver(ActivityRemindedByDriverPushInfo pushInfo) {
+        for (Pusher delegate : delegates) {
+            try {
+                delegate.activityRemindedByDriver(pushInfo);
+            } catch (Exception e) {
+                LOGGER.warn("代理推送器推送消息失败，异常信息如下: ", e);
+            }
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "MultiPusher{" +
+                "delegateTypes='" + delegateTypes + '\'' +
+                ", pusherType='" + pusherType + '\'' +
+                '}';
+    }
+}